+++ Engine/Renderer/OpenGL/GLUtil.h
@@ -6,6 +6,7 @@
 #include "ZeloPrerequisites.h"
 #include "ZeloGLPrerequisites.h"
 #include "Core/RHI/Const/EShaderType.h"
+#include "Core/RHI/Const/EBufferDataType.h"
 
 int checkForOpenGLError(const char *, int);
 
@@ -14,3 +15,5 @@ void initDebugCallback();
 const char *getTypeString(GLenum type);
 
 std::string getShaderTypeString(GLenum shaderType);
+
+GLenum BufferDataTypeToOpenGLBaseType(const Zelo::Core::RHI::EBufferDataType &type);
